Abstract Collection class into a Trait & Interface
